[Fix #262] 죽은 player 처리
